chore: refactor layout to a more structured one
