+++ chapter02/go/begin/appointments/.air.conf
@@ -7,7 +7,7 @@ tmp_dir = "tmp"
 
 [build]
 # Just plain old shell command. You could use `make` as well.
-cmd = "go build -tags dev -o ./appointments ./..."
+cmd = "go build -tags dev -o ./appointments ./main.go"
 # Binary file yields from `cmd`.
 bin = "appointments"
 # Customize binary.
