+++ chapter02/go/begin/appointments/config/config_dev.go
@@ -1,6 +1,6 @@
 //go:build dev
 
-package main
+package config
 
 import (
 	"context"
