+++ chapter02/go/begin/appointments/types/types.go
@@ -1,0 +1,16 @@
+package types
+
+import (
+	"encoding/json"
+	"time"
+)
+
+type Appointment struct {
+	Id              string    `json:"id"`
+	PatientId       string    `json:"patientId"`
+	AppointmentDate time.Time `json:"appointmentDate"`
+}
+
+func (s Appointment) MarshalBinary() ([]byte, error) {
+	return json.Marshal(s)
+}
